:children_crossing: Improve ePub experience
This commit aims at improving the ePub experience for CRS Reports.

It adds a new requirement (PyMuPDF) to handle PDF parsing and
generates HTML from that PDF that is aimed at conversion to ePub.

I also made changes to the build script to add a reports directory
because I couldn't get everything to work locally without it.